Added AssignPassword() function and moved password util functions to utils/password.go
