+++ go.mod
@@ -17,6 +17,7 @@ require (
 	github.com/spf13/cast v1.3.1 // indirect
 	github.com/spf13/cobra v1.1.3
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/spf13/viper v1.7.1
 	github.com/wealdtech/go-merkletree v1.0.0
 	gopkg.in/airbrake/gobrake.v2 v2.0.9 // indirect
